refactory tray out of massive file into smaller one with just new code
