+++ DesktopEdge/Views/Screens/MainMenu.xaml
@@ -53,6 +53,7 @@
                 </Grid>
                 <Rectangle Style="{StaticResource SectionDivider}"/>
                 <StackPanel Name="MainItems" Orientation="Vertical" HorizontalAlignment="Stretch">
+                    <local:OZMenuItem x:Name="ShowWelcomeButton" Label="Show Welcome Screen" IconText="?" MouseUp="ShowWelcomeFromMenu" Visibility="Collapsed"></local:OZMenuItem>
                     <local:OZMenuItem x:Name="IdentitiesButton" Label="Identities" Icon="/Assets/Images/identities.png" MouseUp="ShowIdentities"></local:OZMenuItem>
                     <local:OZMenuItem HorizontalAlignment="Stretch" Width="Auto" MouseUp="ShowAdvanced" Label="Advanced Settings" Icon="/Assets/Images/advanced.png"></local:OZMenuItem>
                     <local:OZMenuItem Label="About" Icon="/Assets/Images/about.png" MouseUp="ShowAbout"></local:OZMenuItem>
